similarity/requirements.txt
python==3.7
django==3.2.13
djangorestframework==3.13.1
django-debug-toolbar==3.8.1
docx==0.2.4
celery
django-celery-beat
django_redis
django-cors-headers==3.13.0
thefuzz==0.19.0
convertapi==1.5.0